refactor: add component Image and use it in PhotoModal and PhotoCard
