+++ src/components/Image/Image.module.css
@@ -1,0 +1,12 @@
+.Image {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  object-fit: contain;
+}
+
+.Visibility {
+  opacity: 0;
+}
