+++ src/components/PhotoCard/PhotoCard.module.css
@@ -30,18 +30,6 @@
   padding-bottom: 80%;
 }
 
-.Image {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  top: 0;
-  left: 0;
-}
-
-.Image_visibility {
-  opacity: 0;
-}
-
 .Title {
   font-weight: 500;
 }
